backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pillow
python-multipart
exifread
python-dotenv
requests

# Civitai automatic authentication

playwright==1.40.0
playwright-stealth